requirements.txt
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coinbase-commerce==1.0.1
colorama==0.4.6
cryptography==44.0.2
Flask==2.1.1
Flask-Login==0.6.2
flask-paypal==0.1.0
Flask-SQLAlchemy==2.5.1
greenlet==3.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
paypalrestsdk==1.14.0
pycparser==2.22
pyOpenSSL==25.0.0
python-dotenv==0.19.2
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.40
stripe==2.72.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3